requirements.txt
arabic-reshaper==3.0.0
asttokens==3.0.0
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
executing==2.1.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
meson==1.6.0
nest-asyncio==1.6.0
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
pybind11==2.13.6
Pygments==2.19.1
python-bidi==0.6.3
pywin32==308
pyzmq==26.2.0
setuptools==75.6.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
wcwidth==0.2.13
accelerate==1.2.1
acres==0.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
diffusers==0.32.1
dill==0.3.8
docker-pycreds==0.4.0
efficientnet_pytorch==0.7.1
einops==0.8.0
etelemetry==0.3.1
evaluate==0.4.3
filelock==3.16.1
fonttools==4.55.3
frontend==0.0.3
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.27.1
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiwer==3.1.0
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
looseversion==1.3.0
lxml==5.3.0
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
munch==4.0.0
natsort==8.4.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.1
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pip==24.3.1
platformdirs==4.3.6
plotly==5.24.1
pretrainedmodels==0.7.4
propcache==0.2.1
protobuf==5.29.3
prov==2.0.1
psutil==6.1.1
puremagic==1.28
pyarrow==19.0.0
pydantic==2.10.5
pydantic_core==2.27.2
pydot==3.0.4
PyMuPDF==1.25.2
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
RapidFuzz==3.12.1
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
segmentation_models_pytorch==0.4.0
sentence-transformers==3.1.1
sentry-sdk==2.19.2
setproctitle==1.3.4
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==1.0.13
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traits==7.0.2
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2025.1
unet==0.8.1
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
